refactor AllPost into MainPost and SubPost
