+++ src/components/Posts/MainPost/index.js
@@ -1,0 +1,1 @@
+export * from './MainPost';
