+++ src/components/Posts/SubPost/index.js
@@ -1,0 +1,1 @@
+export * from './SubPost';
